Add report for Slack's XMPP bridge
Slack provides a XMPP bridge as a kind of freedom-respecting
altiernative to their proprietary Google-Play-only app but it is not
very usable as you can see here.

It does not federate with normal XMPP servers, but I think it still
could make sense to include this report so that people who want to start
using Slack are warned – and maybe use a normal XMPP server instead. ;)

Normally you connect to a subdomain of 'xmpp.slack.com' but I removed it
because this report should be the same on every subdomain of it.